# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "TMM International inc API",
        "description": "Api description...",
        "termsOfService": "",
        "contact": {
            "email": "contact@mysite.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "0.1.0"
    },
    "host": "localhost",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/api/email/test": {
            "post": {
                "tags": [
                    "Emails"
                ],
                "summary": "Send test email",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/email/contact_us": {
            "post": {
                "tags": [
                    "Emails"
                ],
                "summary": "Send contact us info to email",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "FirstName",
                        "in": "formData",
                        "description": "Enter first name",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "string"
                    },
                    {
                        "name": "SecondName",
                        "in": "formData",
                        "description": "Enter second name",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "string"
                    },
                    {
                        "name": "Phone",
                        "in": "formData",
                        "description": "Enter Phone",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "string"
                    },
                    {
                        "name": "Email",
                        "in": "formData",
                        "description": "Enter Email",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "string"
                    },
                    {
                        "name": "Message",
                        "in": "formData",
                        "description": "Enter Message",
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/email/quick_quote": {
            "post": {
                "tags": [
                    "Emails"
                ],
                "summary": "Send QuickQuote to email",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "FreightType",
                        "in": "formData",
                        "description": "Enter first name",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "string"
                    },
                    {
                        "name": "Weight",
                        "in": "formData",
                        "description": "Enter Weight",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "number"
                    },
                    {
                        "name": "ShippingFrom",
                        "in": "formData",
                        "description": "Enter ShippingFrom",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "string"
                    },
                    {
                        "name": "ShippingTo",
                        "in": "formData",
                        "description": "Enter ShippingTo",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "string"
                    },
                    {
                        "name": "Email",
                        "in": "formData",
                        "description": "Enter Email",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "string"
                    },
                    {
                        "name": "Details",
                        "in": "formData",
                        "description": "Enter Details",
                        "schema": {
                            "$ref": "#/definitions/ContactUs"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "definitions": {
        "ContactUs": {
            "properties": {
                "FirstName": {
                    "type": "sting"
                },
                "SecondName": {
                    "type": "sting"
                },
                "Phone": {
                    "type": "sting"
                },
                "Email": {
                    "type": "sting"
                },
                "Message": {
                    "type": "sting"
                }
            }
        },
        "QuickQuote": {
            "properties": {
                "FreightType": {
                    "type": "sting"
                },
                "Weight": {
                    "type": "sting"
                },
                "ShippingFrom": {
                    "type": "sting"
                },
                "ShippingTo": {
                    "type": "sting"
                },
                "Email": {
                    "type": "sting"
                },
                "Details": {
                    "type": "sting"
                }
            }
        },
        "User": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "tags": [
        {
            "name": "Emails"
        }
    ],
    "externalDocs": {
        "description": "Find out more about my website",
        "url": "http..."
    }
}